+++ CHANGELOG.md
@@ -7,6 +7,9 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 
 ## Version 0.6.3 - Unreleased
 
+### Added
+* New CLI script: `coco_move_assets`.
+
 ### Fixed
 * Issue in `coco_subset` CLI when copy-assets is on and the input is absolute paths.
 * reroot can now accept pathlib.Path arguments.